+++ src/commands/hardis/scratch/create.ts
@@ -9,6 +9,7 @@ import * as fs from "fs-extra";
 import * as moment from "moment";
 import * as os from "os";
 import * as path from "path";
+import { clearCache } from "../../../common/cache";
 import { MetadataUtils } from "../../../common/metadata-utils";
 import { elapseEnd, elapseStart, execCommand, execSfdxJson, getCurrentGitBranch, isCI, uxLog } from "../../../common/utils";
 import { deployMetadatas, forceSourceDeploy, forceSourcePush } from "../../../common/utils/deployUtils";
@@ -28,7 +29,18 @@ const messages = Messages.loadMessages("sfdx-hardis", "org");
 export default class ScratchCreate extends SfdxCommand {
   public static title = "Create and initialize scratch org";
 
-  public static description = messages.getMessage("scratchCreate");
+  public static description = `Create and initialize a scratch org or a source-tracked sandbox (config can be defined using \`config/.sfdx-hardis.yml\`):
+
+- **Install packages**
+  - Use property \`installedPackages\`
+- **Push sources**
+- **Assign permission sets**
+  - Use property \`initPermissionSets\`
+- **Run apex initialization scripts**
+  - Use property \`scratchOrgInitApexScripts\`
+- **Load data**
+  - Use property \`dataPackages\`
+  `;
 
   public static examples = ["$ sfdx hardis:scratch:create"];
 
@@ -271,6 +283,7 @@ export default class ScratchCreate extends SfdxCommand {
       output: false,
       debug: this.debugMode,
     });
+    await clearCache("force:org:list");
     assert(createResult.status === 0 && createResult.result, this.buildScratchCreateErrorMessage(createResult));
     this.scratchOrgInfo = createResult.result;
     this.scratchOrgUsername = this.scratchOrgInfo.username;